requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pymysql
redis
pymongo
motor
httpx
aiohttp
python-multipart
fastapi-cors
python-jose
passlib
bcrypt
email-validator
APScheduler
requests
gunicorn